Now generating the correct Comp1 and unComp1s, but still never generating Both.
Also, putting Generic1 in a deriving clause leads to a panic... need to compare
to deriving other (* -> *) classes -- I think that'll be a simple fix.

Otherwise, initial tests indicate that the derivation of Rep1 gdmh.pdf's Figure
3 leads to panics. In particular, the guards of the arg function do not
establish that its last case cannot fire when "the" parameter is in the
type. This error should be caught by cond_functorOK as a "cannot derive
Generic1 for this type" error; instead it leads to a panic because the type
variable occurs free on the RHS of Generic1. Essentially, the "isRep"
conditions need to be moved upstream from arg to cond_functorOK.

If it doesn't panic, it otherwise generates correct code.

Signed-off-by: Jose Pedro Magalhaes <jpm@cs.uu.nl>